+++ core/src/main/java/com/fatpiggies/game/controller/IPlayController.java
@@ -5,5 +5,5 @@ import java.util.ArrayList;
 public interface IPlayController {
     void startGame(String lobbyId, ArrayList<String> playerIds, ArrayList<String> textureIds);
     void endGame(String lobbyId);
-    void updatePlayerInput(int x, int y);
+    void updatePlayerInput(float x, float y);
 }
