+++ core/src/main/java/com/fatpiggies/game/model/GameWorld.java
@@ -60,7 +60,7 @@ public class GameWorld {
      * @param x x direction
      * @param y y direction
      */
-    public void updatePlayerInput(int x, int y) {
+    public void updatePlayerInput(float x, float y) {
         PlayerInputComponent input = localPlayer.getComponent(PlayerInputComponent.class);
         if(input != null) {
             input.joystickPercentageX = x;
